Decouple proxy from filter
All the filtering functions has been implemented in a dedicated class, in this way we can easily implement testing
