+++ go.mod
@@ -2,10 +2,7 @@ module github.com/Ringloop/pisec
 
 go 1.18
 
-require (
-	github.com/bits-and-blooms/bloom/v3 v3.2.0
-	github.com/elazarl/goproxy v0.0.0-20220317163658-f5c0d0953e10
-)
+require github.com/elazarl/goproxy v0.0.0-20220317163658-f5c0d0953e10
 
 require (
 	github.com/cespare/xxhash/v2 v2.1.2 // indirect
@@ -26,6 +23,7 @@ require (
 
 require (
 	github.com/bits-and-blooms/bitset v1.2.0 // indirect
+	github.com/bits-and-blooms/bloom/v3 v3.2.0
 	github.com/go-redis/redis v6.15.9+incompatible
 	github.com/go-redis/redis/v8 v8.11.5
 	github.com/stretchr/testify v1.7.1
